requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
SQLAlchemy==2.0.23
Werkzeug==2.3.7
Jinja2==3.1.2
PyJWT==2.8.0
bleach==6.0.0
Pillow==10.0.1
python-dotenv==1.0.0
email-validator==2.0.0
WTForms==3.0.1
Flask-WTF==1.1.1
click==8.1.3
itsdangerous==2.1.2
six==1.16.0
blinker==1.6.2
alembic==1.12.0
pytz==2023.3 